+++ internal/dataplane/kong_client_test.go
@@ -214,10 +214,12 @@ func (f *mockUpdateStrategyResolver) returnSpecificErrorOnUpdate(url string, err
 	f.errorsToReturnOnUpdate[url] = append(f.errorsToReturnOnUpdate[url], err)
 }
 
+const mockUpdateReturnedConfigSize = 22
+
 // updateCalledForURLCallback returns a function that will be called when the mockUpdateStrategy is called.
 // That enables us to track which URLs were called.
-func (f *mockUpdateStrategyResolver) updateCalledForURLCallback(url string) func(sendconfig.ContentWithHash) error {
-	return func(content sendconfig.ContentWithHash) error {
+func (f *mockUpdateStrategyResolver) updateCalledForURLCallback(url string) func(sendconfig.ContentWithHash) (int, error) {
+	return func(content sendconfig.ContentWithHash) (int, error) {
 		f.lock.Lock()
 		defer f.lock.Unlock()
 
@@ -227,12 +229,11 @@ func (f *mockUpdateStrategyResolver) updateCalledForURLCallback(url string) func
 			if len(errsToReturn) > 0 {
 				err := errsToReturn[0]
 				f.errorsToReturnOnUpdate[url] = errsToReturn[1:]
-				return err
+				return 0, err
 			}
-			return nil
+			return mockUpdateReturnedConfigSize, nil
 		}
-
-		return nil
+		return mockUpdateReturnedConfigSize, nil
 	}
 }
 
@@ -302,10 +303,10 @@ func (f *mockUpdateStrategyResolver) eventuallyGetLastUpdatedContentForURL(
 
 // mockUpdateStrategy is a mock implementation of sendconfig.UpdateStrategy.
 type mockUpdateStrategy struct {
-	onUpdate func(content sendconfig.ContentWithHash) error
+	onUpdate func(content sendconfig.ContentWithHash) (int, error)
 }
 
-func (m *mockUpdateStrategy) Update(_ context.Context, targetContent sendconfig.ContentWithHash) (err error) {
+func (m *mockUpdateStrategy) Update(_ context.Context, targetContent sendconfig.ContentWithHash) (n int, err error) {
 	return m.onUpdate(targetContent)
 }
 
@@ -900,7 +901,7 @@ func TestKongClient_KubernetesEvents(t *testing.T) {
 			}
 			if tc.updateError {
 				if tc.entityErrors {
-					updateStrategyResolver.returnSpecificErrorOnUpdate(testGatewayClient.BaseRootURL(), sendconfig.NewUpdateError(
+					updateStrategyResolver.returnSpecificErrorOnUpdate(testGatewayClient.BaseRootURL(), sendconfig.NewUpdateErrorWithoutResponseBody(
 						[]failures.ResourceFailure{
 							lo.Must(failures.NewResourceFailure("violated constraint", testIngress)),
 							lo.Must(failures.NewResourceFailure("violated constraint", testService)),
@@ -912,7 +913,7 @@ func TestKongClient_KubernetesEvents(t *testing.T) {
 				}
 			}
 			if tc.updateError && tc.fallbackConfigurationUpdateError {
-				updateStrategyResolver.returnSpecificErrorOnUpdate(testGatewayClient.BaseRootURL(), sendconfig.NewUpdateError(
+				updateStrategyResolver.returnSpecificErrorOnUpdate(testGatewayClient.BaseRootURL(), sendconfig.NewUpdateErrorWithoutResponseBody(
 					[]failures.ResourceFailure{
 						lo.Must(failures.NewResourceFailure("violated constraint", testIngress)),
 					}, errors.New("error on update"),
@@ -1170,7 +1171,7 @@ func TestKongClientUpdate_FetchStoreAndPushLastValidConfig(t *testing.T) {
 			},
 		}
 		configBuilder = newMockKongConfigBuilder()
-		updateErr     = sendconfig.NewUpdateError(nil, errors.New("invalid request"))
+		updateErr     = sendconfig.NewUpdateErrorWithoutResponseBody(nil, errors.New("invalid request"))
 	)
 	configBuilder.kongState = newKongState
 
@@ -1361,7 +1362,7 @@ func TestKongClient_FallbackConfiguration_SuccessfulRecovery(t *testing.T) {
 			kongClient.lastValidCacheSnapshot = &lastValidCache
 
 			t.Log("Setting update strategy to return an error on the first call to trigger fallback configuration generation")
-			updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateError(
+			updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateErrorWithoutResponseBody(
 				[]failures.ResourceFailure{
 					lo.Must(failures.NewResourceFailure("violated constraint", brokenConsumer)),
 				},
@@ -1539,7 +1540,7 @@ func TestKongClient_FallbackConfiguration_SkipsUpdateWhenInSync(t *testing.T) {
 		require.NoError(t, originalCache.Add(someConsumer(t, "broken"))) // Add a consumer to cache to change the cache hash.
 
 		t.Log("Setting update strategy to return an error on the first call to trigger fallback configuration generation")
-		updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateError(
+		updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateErrorWithoutResponseBody(
 			[]failures.ResourceFailure{
 				lo.Must(failures.NewResourceFailure("violated constraint", someConsumer(t, "invalid"))),
 			},
@@ -1562,7 +1563,7 @@ func TestKongClient_FallbackConfiguration_SkipsUpdateWhenInSync(t *testing.T) {
 	t.Run("when fallback was used before and config is still broken, after discovering a new client, all clients are updated", func(t *testing.T) {
 		t.Log("Adding a new client to the provider")
 		clientsProvider.gatewayClients = append(clientsProvider.gatewayClients, anotherNewGwClient)
-		updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateError(
+		updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateErrorWithoutResponseBody(
 			[]failures.ResourceFailure{
 				lo.Must(failures.NewResourceFailure("violated constraint", someConsumer(t, "invalid"))),
 			},
@@ -1639,7 +1640,7 @@ func TestKongClient_FallbackConfiguration_FailedRecovery(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Log("Setting update strategy to return an error on the first call to trigger fallback configuration generation")
-	updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateError(
+	updateStrategyResolver.returnSpecificErrorOnUpdate(gwClient.BaseRootURL(), sendconfig.NewUpdateErrorWithoutResponseBody(
 		[]failures.ResourceFailure{
 			lo.Must(failures.NewResourceFailure("violated constraint", brokenConsumer)),
 		},
@@ -1855,7 +1856,7 @@ func TestKongClient_RecoveringFromGatewaySyncError(t *testing.T) {
 		{
 			name: "one of gateways returns UpdateError with entities",
 			errorsFromGateways: []error{
-				sendconfig.NewUpdateError(
+				sendconfig.NewUpdateErrorWithoutResponseBody(
 					[]failures.ResourceFailure{
 						lo.Must(failures.NewResourceFailure("violated constraint", someConsumer(t, "broken"))),
 					},
@@ -1868,7 +1869,7 @@ func TestKongClient_RecoveringFromGatewaySyncError(t *testing.T) {
 		{
 			name: "one of gateways returns UpdateError without entities, has last valid config",
 			errorsFromGateways: []error{
-				sendconfig.NewUpdateError(nil, errors.New("error on update")),
+				sendconfig.NewUpdateErrorWithoutResponseBody(nil, errors.New("error on update")),
 				nil,
 			},
 			hasLastValidConfig:                      true,
@@ -1877,7 +1878,7 @@ func TestKongClient_RecoveringFromGatewaySyncError(t *testing.T) {
 		{
 			name: "one of gateways returns UpdateError without entities, no last valid config",
 			errorsFromGateways: []error{
-				sendconfig.NewUpdateError(nil, errors.New("error on update")),
+				sendconfig.NewUpdateErrorWithoutResponseBody(nil, errors.New("error on update")),
 				nil,
 			},
 			hasLastValidConfig:                       false,
@@ -1897,7 +1898,7 @@ func TestKongClient_RecoveringFromGatewaySyncError(t *testing.T) {
 		{
 			name: "one gateway returns UpdateError, another one an unexpected error",
 			errorsFromGateways: []error{
-				sendconfig.NewUpdateError(
+				sendconfig.NewUpdateErrorWithoutResponseBody(
 					[]failures.ResourceFailure{
 						lo.Must(failures.NewResourceFailure("violated constraint", someConsumer(t, "broken"))),
 					},