+++ internal/dataplane/sendconfig/inmemory.go
@@ -52,17 +52,17 @@ func NewUpdateStrategyInMemory(
 	}
 }
 
-func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState ContentWithHash) error {
+func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState ContentWithHash) (int, error) {
 	dblessConfig := s.configConverter.Convert(targetState.Content)
 	config, err := json.Marshal(dblessConfig)
 	if err != nil {
-		return fmt.Errorf("constructing kong configuration: %w", err)
+		return 0, fmt.Errorf("constructing kong configuration: %w", err)
 	}
 
 	if len(targetState.CustomEntities) > 0 {
 		unmarshaledConfig := map[string]any{}
 		if err := json.Unmarshal(config, &unmarshaledConfig); err != nil {
-			return fmt.Errorf("unmarshaling config for adding custom entities: %w", err)
+			return 0, fmt.Errorf("unmarshaling config for adding custom entities: %w", err)
 		}
 		for entityType, entities := range targetState.CustomEntities {
 			unmarshaledConfig[entityType] = entities
@@ -70,10 +70,11 @@ func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState ContentW
 		}
 		config, err = json.Marshal(unmarshaledConfig)
 		if err != nil {
-			return fmt.Errorf("constructing kong configuration again with custom entities: %w", err)
+			return 0, fmt.Errorf("constructing kong configuration again with custom entities: %w", err)
 		}
 	}
 
+	configSize := len(config)
 	if reloadConfigErr := s.configService.ReloadDeclarativeRawConfig(
 		ctx,
 		bytes.NewReader(config),
@@ -87,23 +88,24 @@ func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState ContentW
 		if errors.As(reloadConfigErr, &apiError) && apiError.Code() == http.StatusBadRequest {
 			resourceErrors, parseErr := parseFlatEntityErrors(apiError.Raw(), s.logger)
 			if parseErr != nil {
-				return fmt.Errorf("failed to parse flat entity errors from error response: %w", parseErr)
+				return 0, fmt.Errorf("failed to parse flat entity errors from error response: %w", parseErr)
 			}
 
 			for _, resourceError := range resourceErrors {
 				s.logger.V(logging.DebugLevel).Info("Resource error", "resource_error", resourceError)
 			}
 
-			return NewUpdateErrorWithResponseBody(
+			return 0, NewUpdateErrorWithResponseBody(
 				apiError.Raw(),
+				configSize,
 				resourceErrorsToResourceFailures(resourceErrors, s.logger),
 				reloadConfigErr,
 			)
 		}
 		// ...otherwise, we return the original one.
-		return fmt.Errorf("failed to reload declarative configuration: %w", reloadConfigErr)
+		return 0, fmt.Errorf("failed to reload declarative configuration: %w", reloadConfigErr)
 	}
-	return nil
+	return configSize, nil
 }
 
 func (s UpdateStrategyInMemory) MetricsProtocol() metrics.Protocol {