+++ internal/metrics/prometheus_test.go
@@ -25,27 +25,29 @@ func TestNewCtrlFuncMetricsDoesNotPanicWhenCalledTwice(t *testing.T) {
 }
 
 func TestRecordPush(t *testing.T) {
+	const mockSizeOfCfg = 22
 	m := NewCtrlFuncMetrics()
+
 	t.Run("recording push success works", func(t *testing.T) {
 		require.NotPanics(t, func() {
-			m.RecordPushSuccess(ProtocolDBLess, time.Millisecond, "https://10.0.0.1:8080")
+			m.RecordPushSuccess(ProtocolDBLess, time.Millisecond, mockSizeOfCfg, "https://10.0.0.1:8080")
 		})
 	})
 	t.Run("recording push failure works", func(t *testing.T) {
 		require.NotPanics(t, func() {
-			m.RecordPushFailure(ProtocolDBLess, time.Millisecond, "https://10.0.0.1:8080", 5, fmt.Errorf("custom error"))
+			m.RecordPushFailure(ProtocolDBLess, time.Millisecond, mockSizeOfCfg, "https://10.0.0.1:8080", 5, fmt.Errorf("custom error"))
 		})
 	})
 	// Verify that multiple call of NewCtrlFuncMetrics keeps all created metrics work.
 	m2 := NewCtrlFuncMetrics()
 	t.Run("recording push success works for old metrics", func(t *testing.T) {
 		require.NotPanics(t, func() {
-			m.RecordPushSuccess(ProtocolDBLess, time.Millisecond, "https://10.0.0.1:8080")
+			m.RecordPushSuccess(ProtocolDBLess, time.Millisecond, mockSizeOfCfg, "https://10.0.0.1:8080")
 		})
 	})
 	t.Run("recording push success works for new metrics", func(t *testing.T) {
 		require.NotPanics(t, func() {
-			m2.RecordPushSuccess(ProtocolDBLess, time.Millisecond, "https://10.0.0.2:8080")
+			m2.RecordPushSuccess(ProtocolDBLess, time.Millisecond, mockSizeOfCfg, "https://10.0.0.2:8080")
 		})
 	})
 }