+++ test/envtest/metrics_envtest_test.go
@@ -48,12 +48,14 @@ func TestMetricsAreServed(t *testing.T) {
 				metrics.MetricNameConfigPushBrokenResources,
 				metrics.MetricNameTranslationCount,
 				metrics.MetricNameTranslationDuration,
+				metrics.MetricNameConfigPushSize,
 				metrics.MetricNameTranslationBrokenResources,
 				metrics.MetricNameConfigPushDuration,
 
 				metrics.MetricNameFallbackTranslationBrokenResources,
 				metrics.MetricNameFallbackTranslationCount,
 				metrics.MetricNameFallbackTranslationDuration,
+				metrics.MetricNameFallbackConfigPushSize,
 				metrics.MetricNameFallbackConfigPushCount,
 				metrics.MetricNameFallbackConfigPushSuccessTime,
 				metrics.MetricNameFallbackConfigPushDuration,
@@ -70,6 +72,7 @@ func TestMetricsAreServed(t *testing.T) {
 				metrics.MetricNameConfigPushBrokenResources,
 				metrics.MetricNameTranslationCount,
 				metrics.MetricNameTranslationDuration,
+				metrics.MetricNameConfigPushSize,
 				metrics.MetricNameTranslationBrokenResources,
 				metrics.MetricNameConfigPushDuration,
 				metrics.MetricNameConfigPushSuccessTime,
@@ -99,12 +102,10 @@ func TestMetricsAreServed(t *testing.T) {
 				},
 			)
 
-			wantMetrics := tc.expectedMetrics
-
 			metricsURL := fmt.Sprintf("http://%s/metrics", cfg.MetricsAddr)
 			t.Logf("waiting for metrics to be available at %q", metricsURL)
 
-			for _, metric := range wantMetrics {
+			for _, metric := range tc.expectedMetrics {
 				t.Run(metric, func(t *testing.T) {
 					require.NoError(t,
 						retry.Do(func() error {
@@ -112,8 +113,8 @@ func TestMetricsAreServed(t *testing.T) {
 							if err != nil {
 								return fmt.Errorf("error %w checking %q", err, metricsURL)
 							}
-
 							defer resp.Body.Close()
+
 							if http.StatusOK != resp.StatusCode {
 								return fmt.Errorf("status code %v not as expected (200)", resp.StatusCode)
 							}